Upgrade oauth2 to current version.
The 0.9.x version of oauth2 still rely on the default SSL settings. As a result, it attempts to connect to the authorization server using SSLv3. It appears that your authorization server no longer support that protocol (which is a good thing.).

Your tests are still passing after upgrading oauth2 to 1.1.0. This should address #13 .